#VER-72 - Algumas melhorias de pedidos e trasnferencias

+++ VersopayBackend/Dtos/VexyBank/PixInCreateReqDto.cs
@@ -4,7 +4,7 @@ namespace VersopayBackend.Dtos.VexyBank
 {
     public sealed class PixInCreateReqDto
     {
-        [JsonPropertyName("amountInCents")] public int AmountInCents { get; set; }
+        [JsonPropertyName("amountInCents")] public decimal AmountInCents { get; set; }
         [JsonPropertyName("description")] public string? Description { get; set; }
         [JsonPropertyName("postbackUrl")] public string? PostbackUrl { get; set; }
         [JsonPropertyName("customer")] public PixCustomerDto Customer { get; set; } = new();
